requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
djangorestframework==3.10.3
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==20.0.2
h5py==2.10.0
idna==2.8
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
nltk==3.4.5
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
protobuf==3.10.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.3
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==2.0.1
termcolor==1.1.0
textblob==0.15.3
tweepy==3.8.0
urllib3==1.25.7
Werkzeug==0.16.0
whitenoise==4.1.4
wrapt==1.11.2
